Moved files in table folder
